more things to pairlistStore
